requirements.txt
fastapi==0.115.*
uvicorn==0.34.*
aiosqlite==0.21.*
apscheduler==3.11.*
firebase-admin==6.6.*
python-dotenv==1.1.*
